requirements.txt
###### Requirements without Version Specifiers ######`
click
pandas
python_dateutil
click_logging
monthdelta
